+++ frm/components/Navigation/NavigationSlotsGroupsBreakpoints.js
@@ -24,7 +24,6 @@ export default /* html */`
     <button
       class="nav-hide"
       type="button"
-      aria-haspopup="true"
       data-nav-open
     >
       Open