+++ frm/objects/Media/MediaPartial.js
@@ -1,0 +1,12 @@
+export default /* html */`
+  <frm-media
+    id="med-partial"
+    class="block"
+    title="Partial video title"
+    loader="med-loader"
+    error="med-error-none"
+  >
+    <video class="ar-16-9 w-full"></video>
+    <button type="button" data-media-control="toggle">Toggle</button>
+  </frm-media>
+`