+++ src/components/Navigation/__tests__/Navigation.spec.ts
@@ -163,7 +163,6 @@ test.describe('Navigation', () => {
         modalRole: nav.modal?.role,
         modalSlotsSize: nav.modalSlots.size,
         breakpoints: Array.from(nav.breakpoints.keys()).join(','),
-        opensPopup: nav.opens?.ariaHasPopup,
         closesCount: nav.closes.length
       }
     })
@@ -174,13 +173,20 @@ test.describe('Navigation', () => {
     expect(navProps.modalRole).toBe('dialog')
     expect(navProps.modalSlotsSize).toBe(1)
     expect(navProps.breakpoints).toBe('0')
-    expect(navProps.opensPopup).toBe('true')
     expect(navProps.closesCount).toBe(1)
   })
 
   /* Test resize */
 
   test('should move some items into modal slots if 900px viewport', async ({ page }) => {
+    await page.evaluate(async () => { // Resize flag
+      const { onResize } = await import('../../../actions/actionResize.js')
+
+      onResize(() => {
+        window.testNavResize = true
+      })
+    })
+
     const viewport = page.viewportSize() as { width: number, height: number }
 
     await page.setViewportSize({
@@ -189,7 +195,7 @@ test.describe('Navigation', () => {
     })
 
     await page.waitForFunction(() => { // Wait for resize
-      return window.testNavSet.filter(id => id === 'nav-slots-groups-breakpoints').length === 2
+      return window.testNavResize
     })
 
     const navSlots = await page.evaluate(() => {
@@ -217,6 +223,14 @@ test.describe('Navigation', () => {
   })
 
   test('should move all items into modal slots if 600px viewport', async ({ page }) => {
+    await page.evaluate(async () => { // Resize flag
+      const { onResize } = await import('../../../actions/actionResize.js')
+
+      onResize(() => {
+        window.testNavResize = true
+      })
+    })
+
     const viewport = page.viewportSize() as { width: number, height: number }
 
     await page.setViewportSize({
@@ -225,7 +239,7 @@ test.describe('Navigation', () => {
     })
 
     await page.waitForFunction(() => { // Wait for resize
-      return window.testNavSet.filter(id => id === 'nav-slots-groups-breakpoints').length === 2
+      return window.testNavResize
     })
 
     const navSlots = await page.evaluate(() => {
@@ -266,7 +280,7 @@ test.describe('Navigation', () => {
   })
 
   test('should not reset if viewport height change', async ({ page }) => {
-    await page.evaluate(async () => {
+    await page.evaluate(async () => { // Resize flag
       const { onResize } = await import('../../../actions/actionResize.js')
 
       onResize(() => {
@@ -307,6 +321,14 @@ test.describe('Navigation', () => {
   })
 
   test('should not overflow if slots undefined', async ({ page }) => {
+    await page.evaluate(async () => { // Resize flag
+      const { onResize } = await import('../../../actions/actionResize.js')
+
+      onResize(() => {
+        window.testNavResize = true
+      })
+    })
+
     await page.evaluate(() => {
       const nav = document.querySelector('#nav-slot') as Navigation
       nav.slots.clear()
@@ -320,7 +342,7 @@ test.describe('Navigation', () => {
     })
 
     await page.waitForFunction(() => { // Wait for resize
-      return window.testNavSet.filter(id => id === 'nav-slot').length === 2
+      return window.testNavResize
     })
 
     const navProps = await page.evaluate(() => {
@@ -339,6 +361,14 @@ test.describe('Navigation', () => {
   /* Test modal */
 
   test('should open and close modal', async ({ page }) => {
+    await page.evaluate(async () => { // Resize flag
+      const { onResize } = await import('../../../actions/actionResize.js')
+
+      onResize(() => {
+        window.testNavResize = true
+      })
+    })
+
     const viewport = page.viewportSize() as { width: number, height: number }
 
     await page.setViewportSize({
@@ -347,7 +377,7 @@ test.describe('Navigation', () => {
     })
 
     await page.waitForFunction(() => { // Wait for resize
-      return window.testNavSet.filter(id => id === 'nav-slot').length === 2
+      return window.testNavResize
     })
 
     await page.getByTestId('nav-slot-open').click()
@@ -417,6 +447,14 @@ test.describe('Navigation', () => {
   })
 
   test('should close modal on escape', async ({ page }) => {
+    await page.evaluate(async () => { // Resize flag
+      const { onResize } = await import('../../../actions/actionResize.js')
+
+      onResize(() => {
+        window.testNavResize = true
+      })
+    })
+
     const viewport = page.viewportSize() as { width: number, height: number }
 
     await page.setViewportSize({
@@ -425,7 +463,7 @@ test.describe('Navigation', () => {
     })
 
     await page.waitForFunction(() => { // Wait for resize
-      return window.testNavSet.filter(id => id === 'nav-slots').length === 2
+      return window.testNavResize
     })
 
     await page.getByTestId('nav-slots-open').click()
@@ -457,6 +495,14 @@ test.describe('Navigation', () => {
   })
 
   test('should close modal on resize', async ({ page }) => {
+    await page.evaluate(async () => { // Resize flag
+      const { onResize } = await import('../../../actions/actionResize.js')
+
+      onResize(() => {
+        window.testNavResize = true
+      })
+    })
+
     const viewport = page.viewportSize() as { width: number, height: number }
 
     await page.setViewportSize({
@@ -465,7 +511,7 @@ test.describe('Navigation', () => {
     })
 
     await page.waitForFunction(() => { // Wait for resize
-      return window.testNavSet.filter(id => id === 'nav-slot').length === 2
+      return window.testNavResize
     })
 
     await page.getByTestId('nav-slot-open').click()
@@ -482,7 +528,7 @@ test.describe('Navigation', () => {
     })
 
     await page.waitForFunction(() => { // Wait for resize
-      return window.testNavSet.filter(id => id === 'nav-slot').length === 3
+      return window.testNavResize
     })
 
     await page.waitForFunction(() => { // Wait for close