+++ CHANGELOG.md
@@ -5,3 +5,16 @@ All notable changes to this project will be documented in this file.
 ## [0.0.1] - 2026-02-13
 
 - Initial release.
+
+## [0.0.2] - 2026-04-26
+
+### Changed
+
+- Update loader z-index to CSS variable.
+- `Media` replace touch and mouse with pointer event listeners.
+
+### Fixed
+
+- `Media` duration text, load and mute on progress interaction.
+- `getDuration` returning empty string for 0 seconds when words true. 
+- z-index utility class name.